+++ app/build.gradle
@@ -44,6 +44,7 @@ android {
         implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
         implementation("androidx.navigation:navigation-fragment-ktx:$nav_version")
         implementation("androidx.navigation:navigation-ui-ktx:$nav_version")
+        implementation("androidx.lifecycle:lifecycle-viewmodel-ktx:$lifecycle_version")
         testImplementation 'junit:junit:4.13.2'
         androidTestImplementation 'androidx.test.ext:junit:1.1.3'
         androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
