+++ app/src/main/java/com/azimut4946777/guessinggame/GameFragment.kt
@@ -5,19 +5,15 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.findNavController
 import com.azimut4946777.guessinggame.databinding.FragmentGameBinding
 
 class GameFragment : Fragment() {
     private var _binding: FragmentGameBinding? = null
     private val binding get() = _binding!!
+   lateinit var viewModel: GameViewModel
 
-    val words = listOf("Android,", "Activity", "Fragment")
-    val secretWord = words.random().uppercase()
-    var secretWordDisplay = ""
-    var correctGuesses = ""
-    var incorrectGuesses = ""
-    var livesLeft = 8
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -25,16 +21,16 @@ class GameFragment : Fragment() {
     ): View? {
         _binding = FragmentGameBinding.inflate(inflater, container, false)
         val view = binding.root
-        secretWordDisplay = deriveSecretWordDisplay()
+viewModel = ViewModelProvider(this).get(GameViewModel::class.java)
         updateScreen()
 
         binding.guessButton.setOnClickListener {
-            makeGuesses(binding.guess.text.toString().uppercase())
+            viewModel.makeGuesses(binding.guess.text.toString().uppercase())
             binding.guess.text = null
             updateScreen()
-            if(isWon()||isLost()) {
+            if(viewModel.isWon()||viewModel.isLost()) {
                 val action = GameFragmentDirections.
-                        actionGameFragmentToResultFragment(wonLostMessage())
+                        actionGameFragmentToResultFragment(viewModel.wonLostMessage())
                 view.findNavController().navigate(action)
             }
         }
@@ -49,46 +45,11 @@ class GameFragment : Fragment() {
     }
 
     fun updateScreen() {
-        binding.word.text = secretWordDisplay
-        binding.lives.text = "You Have $livesLeft Lives Left"
-        binding.incorrectGuesses.text = "Incorrect Guesses: $incorrectGuesses"
+        binding.word.text = viewModel.secretWordDisplay
+        binding.lives.text = "You Have ${viewModel.livesLeft} Lives Left"
+        binding.incorrectGuesses.text = "Incorrect Guesses: ${viewModel.incorrectGuesses}"
     }
 
-    fun deriveSecretWordDisplay(): String {
-        var display = ""
-        secretWord.forEach {
-            display += checkLetter(it.toString())
-        }
-        return display
-    }
-
-    fun checkLetter(str: String) = when (correctGuesses.contains(str)) {
-        true -> str
-        false -> "_"
-    }
-
-    fun makeGuesses(guess: String) {
-        if (guess.length == 1) {
-            if (secretWord.contains(guess)) {
-                correctGuesses += guess
-                secretWordDisplay = deriveSecretWordDisplay()
-            } else {
-                incorrectGuesses += "$guess"
-                livesLeft--
-            }
-        }
-    }
-
-    fun isWon() = secretWord.equals(secretWordDisplay, true)
-    fun isLost() = livesLeft <= 0
-
-    fun wonLostMessage(): String {
-        var message = ""
-        if (isWon()) message = "You Won!"
-        else if (isLost()) message = "You Lost!"
-        message += "The word was $secretWord"
-        return message
-    }
 
 
 }
