+++ app/src/main/java/com/azimut4946777/guessinggame/GameViewModel.kt
@@ -1,0 +1,54 @@
+package com.azimut4946777.guessinggame
+
+import androidx.lifecycle.ViewModel
+
+class GameViewModel: ViewModel() {
+
+    val words = listOf("Android,", "Activity", "Fragment")
+    val secretWord = words.random().uppercase()
+    var secretWordDisplay = ""
+    var correctGuesses = ""
+    var incorrectGuesses = ""
+    var livesLeft = 8
+    init {
+        secretWordDisplay = deriveSecretWordDisplay()
+
+    }
+    fun deriveSecretWordDisplay(): String {
+        var display = ""
+        secretWord.forEach {
+            display += checkLetter(it.toString())
+        }
+        return display
+    }
+
+    fun checkLetter(str: String) = when (correctGuesses.contains(str)) {
+        true -> str
+        false -> "_"
+    }
+    fun makeGuesses(guess: String) {
+        if (guess.length == 1) {
+            if (secretWord.contains(guess)) {
+                correctGuesses += guess
+                secretWordDisplay = deriveSecretWordDisplay()
+            } else {
+                incorrectGuesses += "$guess"
+                livesLeft--
+            }
+        }
+    }
+
+    fun isWon() = secretWord.equals(secretWordDisplay, true)
+    fun isLost() = livesLeft <= 0
+
+    fun wonLostMessage(): String {
+        var message = ""
+        if (isWon()) message = "You Won!"
+        else if (isLost()) message = "You Lost!"
+        message += "The word was $secretWord"
+        return message
+    }
+
+
+}
+
