+++ build.gradle
@@ -2,6 +2,7 @@
 buildscript {
 
     ext.nav_version = '2.5.0'
+    ext.lifecycle_version = "2.6.0-alpha01"
     repositories {
         google()
         jcenter()
